RavenDB-19633 cleanup
